# openapi.yaml
openapi: '3.1.0'
info:
  title: 'Mission Control API'
  version: '1.0.0'
  description: |-
    Generated from the Backend’s Fastify route table by `pnpm api:spec` (F5.1). Do not edit
    by hand: `pnpm api:spec:check` fails when this file and the routes disagree, and
    `pnpm api:types:check` fails when apps/frontend/src/lib/api/types.generated.ts does.

    REQUEST shapes are the schemas Ajv actually enforces, so they are exact.

    RESPONSE shapes are the schemas each route declares under `schema.response`. Read the
    guarantee precisely: they are DECLARATIONS, NOT SERIALIZERS. The Backend installs a
    serializer compiler that ignores them and calls JSON.stringify, so a schema here can never
    remove a field the handler produced — the failure mode a Fastify response schema normally
    carries. What keeps them true instead is a conformance check that validates every reply
    against its own schema in the test tiers, in both directions: an undeclared property and a
    missing required property are both failures. See apps/backend/src/http/response-schema.ts.

    Operations still marked `x-mc-response-schema: undeclared` are the ones with no declaration,
    and the marker exists so that gap stays countable. Today they are: GET /health,
    GET /services/health and GET /schedule (owned by modules outside this change),
    GET|PUT /settings/{category} (the GET serves one of five different documents, which needs a
...
    and GET /ws.

...
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
  /adrs:
    get:
      operationId: getAdrs
      summary: 'GET /adrs'
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - proposed
              - accepted
              - rejected
              - superseded
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Adr'
                  meta:
                    $ref: '#/components/schemas/ListEnvelopeMeta'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    post:
      operationId: postAdrs
      summary: 'POST /adrs'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - title
              additionalProperties: false
              properties:
                projectId:
                  type: string
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                title:
                  type: string
                  minLength: 1
                  maxLength: 300
                status:
                  type: string
                  enum:
                    - proposed
                    - accepted
                    - rejected
                    - superseded
                context:
                  type: string
                  maxLength: 100000
                decision:
                  type: string
                  maxLength: 100000
                alternatives:
                  type: string
                  maxLength: 100000
                consequences:
                  type: string
                  maxLength: 100000
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Adr'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  '/adrs/{id}':
    get:
      operationId: getAdrsById
      summary: 'GET /adrs/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Adr'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    patch:
      operationId: patchAdrsById
      summary: 'PATCH /adrs/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 300
                status:
                  type: string
                  enum:
                    - proposed
                    - accepted
                    - rejected
                    - superseded
                context:
                  type: string
                  maxLength: 100000
                decision:
                  type: string
                  maxLength: 100000
                alternatives:
                  type: string
                  maxLength: 100000
                consequences:
                  type: string
                  maxLength: 100000
                supersededByAdrId:
                  type:
                    - string
                    - 'null'
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Adr'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
  /sessions:
    get:
      operationId: getSessions
      summary: 'GET /sessions'
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        - name: repositoryId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        - name: sessionType
          in: query
          required: false
          schema:
            type: string
            enum:
              - managed
              - observed
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum:
              - created
              - running
              - paused
              - completed
              - failed
              - archived
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  meta:
                    $ref: '#/components/schemas/ListEnvelopeMeta'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    post:
      operationId: postSessions
      summary: 'POST /sessions'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - workingDirectory
              additionalProperties: false
              properties:
                projectId:
                  type: string
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                workingDirectory:
                  type: string
                  minLength: 1
                  maxLength: 4096
                repositoryId:
                  type: string
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                branch:
                  type: string
                  maxLength: 255
                title:
                  type: string
                  maxLength: 200
                model:
                  type: string
                  maxLength: 128
                agentId:
                  type: string
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Session'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
  '/sessions/{id}/generate-adr':
    post:
      operationId: postSessionsByIdGenerateAdr
      summary: 'POST /sessions/{id}/generate-adr'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '202':
          description: 'Accepted — the work is queued; the outcome is observable on the resource it changes.'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/JobAccepted'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
components:
  schemas:
    Adr:
      type: object
      additionalProperties: false
      required:
        - id
        - projectId
        - adrNumber
        - title
        - status
        - context
        - decision
        - alternatives
        - consequences
        - sourceSessionId
        - supersededByAdrId
        - obsidianPath
        - syncedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        projectId:
          type: string
        adrNumber:
          type: integer
        title:
          type: string
        status:
          $ref: '#/components/schemas/AdrStatus'
        context:
          type: string
        decision:
          type: string
        alternatives:
          type: string
        consequences:
          type: string
        sourceSessionId:
          type:
            - string
            - 'null'
        supersededByAdrId:
          type:
            - string
            - 'null'
        obsidianPath:
          type:
            - string
            - 'null'
        syncedAt:
          type:
            - string
            - 'null'
        createdAt:
          type: string
        updatedAt:
          type: string
    AdrStatus:
      type: string
      enum:
        - proposed
        - accepted
        - rejected
        - superseded
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
            - details
            - requestId
          properties:
            code:
              $ref: '#/components/schemas/ErrorEnvelopeCode'
            message:
              type: string
            details:
              type:
                - object
                - 'null'
            requestId:
              type: string
              description: 'Also returned as the X-Request-Id header, and in every log line.'
    ErrorEnvelopeCode:
      type: string
      description: |-
        The F5.4 error-code registry (TDS 04 §1.3), verbatim from the Backend’s http/errors.ts.
        Named ErrorEnvelopeCode rather than ErrorCode because the SPA already owns that name for a
        deliberately wider union: lib/api/errors.ts admits codes the client synthesises for failures
        that never reached the server (NETWORK_ERROR, ABORTED). This is the server’s list only.
      enum:
        - CONFLICT
        - DATABASE_SCHEMA_MISMATCH
        - FORBIDDEN
        - INTEGRATION_NOT_CONFIGURED
        - INTERNAL
        - INVALID_CREDENTIALS
        - INVALID_CURSOR
        - INVALID_STATE_TRANSITION
        - NOT_FOUND
        - NO_TURN_IN_FLIGHT
        - OPERATION_NOT_SUPPORTED
        - ORIGIN_NOT_ALLOWED
        - PAYLOAD_TOO_LARGE
        - RATE_LIMITED
        - RUNTIME_UNAVAILABLE
        - SESSION_NOT_RUNNING
        - UNAUTHORIZED
        - VALIDATION_FAILED
...
    JobAccepted:
      type: object
      additionalProperties: false
      required:
        - jobId
      properties:
        jobId:
          type: string
...
    ListEnvelopeMeta:
      type: object
      additionalProperties: false
      required:
        - nextCursor
        - limit
      properties:
        nextCursor:
          type:
            - string
            - 'null'
        limit:
          type: integer
...
    Session:
      type: object
      additionalProperties: false
      required:
        - id
        - projectId
        - repositoryId
        - sessionType
        - state
        - title
        - failureReason
        - notes
        - branch
        - workingDirectory
        - agentId
        - runtime
        - observation
        - costUsd
        - tokenUsage
        - durationSeconds
        - resumedFromSessionId
        - clonedFromSessionId
        - createdAt
        - startedAt
        - completedAt
        - archivedAt
        - updatedAt
      properties:
        id:
          type: string
        projectId:
          type: string
        repositoryId:
          type:
            - string
            - 'null'
        sessionType:
          $ref: '#/components/schemas/SessionType'
        state:
          $ref: '#/components/schemas/SessionState'
        title:
          type: string
          description: 'Always a string; '''' is the unset value, not a title of zero length. Display falls back to a derived label.'
        failureReason:
          type:
            - string
            - 'null'
          description: 'Why a `failed` Session failed - spawn_error, process_crash, backend_restart. `null` in every other state, and legitimately `null` in `failed` too when the transition carried no reason. It is the whole triage signal.'
        notes:
          type:
            - string
            - 'null'
        branch:
          type:
            - string
            - 'null'
        workingDirectory:
          type: string
        agentId:
          type:
            - string
            - 'null'
          description: 'The Agent persona this Session runs as (PRD 5.1), or null. The id ONLY - the Agent resource is served by GET /agents/{id}, because inlining a persona would put a 20 000-character system prompt on the session list.'
        runtime:
          $ref: '#/components/schemas/SessionRuntimeInfo'
        observation:
          anyOf:
            - $ref: '#/components/schemas/SessionObservation'
            - type: 'null'
        costUsd:
          type:
            - number
            - 'null'
        tokenUsage:
          anyOf:
            - $ref: '#/components/schemas/SessionTokenUsage'
            - type: 'null'
        durationSeconds:
          type:
            - integer
            - 'null'
        resumedFromSessionId:
          type:
            - string
            - 'null'
        clonedFromSessionId:
          type:
            - string
            - 'null'
        createdAt:
          type: string
        startedAt:
          type:
            - string
            - 'null'
        completedAt:
          type:
            - string
            - 'null'
        archivedAt:
          type:
            - string
            - 'null'
        updatedAt:
          type: string
...
    SessionObservation:
      type: object
      additionalProperties: false
      required:
        - channel
        - degraded
        - reason
        - driftCount
        - updatedAt
      properties:
        channel:
          type: string
          enum:
            - hooks_and_transcript
            - hooks_only
            - transcript_only
        degraded:
          type: boolean
        reason:
          type:
            - string
            - 'null'
        driftCount:
          type: integer
        updatedAt:
          type: string
    SessionRuntimeInfo:
      type: object
      additionalProperties: false
      required:
        - kind
        - runtimeSessionId
        - claudeVersion
        - model
        - machine
        - environment
      properties:
        kind:
          type: string
        runtimeSessionId:
          type:
            - string
            - 'null'
        claudeVersion:
          type:
            - string
            - 'null'
        model:
          type:
            - string
            - 'null'
        machine:
          type:
            - string
            - 'null'
        environment:
          type:
            - string
            - 'null'
    SessionState:
      type: string
      enum:
        - created
        - running
        - paused
        - completed
        - failed
        - archived
    SessionTokenUsage:
      type: object
      additionalProperties: false
      required:
        - input
        - output
        - cacheRead
        - cacheWrite
      properties:
        input:
          type: integer
        output:
          type: integer
        cacheRead:
          type: integer
        cacheWrite:
          type: integer
    SessionType:
      type: string
      enum:
        - managed
        - observed
...
  responses:
    ErrorEnvelope:
      description: 'Any non-2xx response (F5.4). The status code and the `error.code` registry entry carry the meaning; see docs/tds/04-api-contracts-and-events.md §1.3.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mc_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'mct_ API token'